force fp16

+++ minimal_multitask/nn_influence_utils.py
@@ -466,7 +466,7 @@ def compute_influences_train_index(
     influences = {}
 
     # flatten s_test
-    s_test = torch.cat([g.reshape(-1) for g in s_test], axis=0)
+    s_test = torch.cat([g.reshape(-1) for g in s_test], axis=0).to(torch.float16)
     # if we are using a projector, project the s_test
     if projector is not None:
         s_test = projector.project(s_test.view(1, -1), model_id=0)
